Add comments and move helper classes to TestUtils
