+++ TestUtils/src/main/java/com/microsoft/device/dualscreen/testutils/FoldHelper.kt
@@ -6,6 +6,7 @@
 package com.microsoft.device.dualscreen.testutils
 
 import androidx.activity.ComponentActivity
+import androidx.compose.ui.test.junit4.AndroidComposeTestRule
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.window.layout.FoldingFeature
 import androidx.window.testing.layout.FoldingFeature
@@ -80,4 +81,57 @@ fun <A : ComponentActivity> WindowLayoutInfoPublisherRule.simulateFold(
         val windowLayoutInfo = TestWindowLayoutInfo(listOf(fold))
         overrideWindowLayoutInfo(windowLayoutInfo)
     }
+}
+
+/**
+ * Simulate a vertical fold in a Compose test
+ *
+ * @param composeTestRule: Compose android test rule
+ * @param center: location of center of fold
+ * @param size: size of fold
+ * @param state: state of fold
+ */
+fun <A : ComponentActivity> WindowLayoutInfoPublisherRule.simulateVerticalFold(
+    composeTestRule: AndroidComposeTestRule<ActivityScenarioRule<A>, A>,
+    center: Int = -1,
+    size: Int = 0,
+    state: FoldingFeature.State = FoldingFeature.State.HALF_OPENED
+) {
+    simulateVerticalFold(composeTestRule.activityRule, center, size, state)
+}
+
+/**
+ * Simulate a horizontal fold in a Compose test
+ *
+ * @param composeTestRule: Compose android test rule
+ * @param center: location of center of fold
+ * @param size: size of fold
+ * @param state: state of fold
+ */
+fun <A : ComponentActivity> WindowLayoutInfoPublisherRule.simulateHorizontalFold(
+    composeTestRule: AndroidComposeTestRule<ActivityScenarioRule<A>, A>,
+    center: Int = -1,
+    size: Int = 0,
+    state: FoldingFeature.State = FoldingFeature.State.HALF_OPENED
+) {
+    simulateHorizontalFold(composeTestRule.activityRule, center, size, state)
+}
+
+/**
+ * Simulate a fold with the given properties in a Compose test
+ *
+ * @param composeTestRule: Compose android test rule
+ * @param center: location of center of fold
+ * @param size: size of fold
+ * @param state: state of fold
+ * @param orientation: orientation of fold
+ */
+fun <A : ComponentActivity> WindowLayoutInfoPublisherRule.simulateFold(
+    composeTestRule: AndroidComposeTestRule<ActivityScenarioRule<A>, A>,
+    center: Int,
+    size: Int,
+    state: FoldingFeature.State,
+    orientation: FoldingFeature.Orientation,
+) {
+    simulateFold(composeTestRule.activityRule, center, size, state, orientation)
 }
