+++ TestUtils/src/main/java/com/microsoft/device/dualscreen/testutils/StringHelper.kt
@@ -1,0 +1,26 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+package com.microsoft.device.dualscreen.testutils
+
+import androidx.activity.ComponentActivity
+import androidx.annotation.StringRes
+import androidx.compose.ui.test.junit4.AndroidComposeTestRule
+import androidx.test.ext.junit.rules.ActivityScenarioRule
+
+/**
+ * STRING HELPER
+ * -----------------------------------------------------------------------------------------------
+ * These functions can be used for string operations in UI tests to simplify testing code.
+ */
+
+/**
+ * Get resource string inside Compose test with resource id
+ *
+ * @param id: string resource id
+ */
+fun <A : ComponentActivity> AndroidComposeTestRule<ActivityScenarioRule<A>, A>.getString(@StringRes id: Int): String {
+    return activity.getString(id)
+}
